Revert "Import Cronet 7b194e46 (139.0.7216.0) into tot"
This reverts commit 8739c9f788d46f94122b89666c3c2711276e64f4.

Reason for revert: Likely culprit for b/421867066  - verifying through ABTD before revert submission. This is part of the standard investigation process, and does not mean your CL will be reverted.


Bug : 421867066

+++ tot/net/log/net_log_capture_mode.h
@@ -21,9 +21,7 @@ namespace net {
 //
 // Note the numeric values are used in a bitfield (NetLogCaptureModeSet) so must
 // be sequential starting from 0, and not exceed 31.
-//
-// GENERATED_JAVA_ENUM_PACKAGE: org.chromium.net
-enum class NetLogCaptureMode : uint8_t {
+enum class NetLogCaptureMode : uint32_t {
   // Logging level that only allows fields that are known to be safe from a
   // privacy perspective.
   //
@@ -68,11 +66,11 @@ enum class NetLogCaptureMode : uint8_t {
 // value "i".
 //
 // Use the NetLogCaptureModeSet*() functions to operate on it.
-using NetLogCaptureModeSet = uint8_t;
+using NetLogCaptureModeSet = uint32_t;
 
 inline NetLogCaptureModeSet NetLogCaptureModeToBit(
     NetLogCaptureMode capture_mode) {
-  return static_cast<uint8_t>(1 << static_cast<uint8_t>(capture_mode));
+  return 1 << static_cast<uint32_t>(capture_mode);
 }
 
 inline bool NetLogCaptureModeSetContains(NetLogCaptureMode capture_mode,
